feat: added pagination to gitlab commit collector
